- Added new iterative projection algorithm which uses robust techniques - Added functions to create/maintain the directories in which all data/images will now be organized. - Added struct variable whose members are associated with the preprocessing/reconstruction parameters so users can define and control program behavior through key/value formatted configuration files.
Includes 2 implementations of robust approach which differ in when it is applied, either during calculation or application of image update.

The program is to be separated into preprocessing and reconstruction stages so they can be ran consecutively or as individual modules, and this black box approach requires function and input/output characteristics to be standardized.  Converting user defined options from programmatic to external control through configuration files in key/value format, users need not learn the details of the implementation and the program can be executed with different parameters without recompiling.

As these parameters will no longer be known at compile time, these can no longer be stored in const/#define global variables, which means all parameters relevant to GPU computation must be passed to the GPU explicitly on each kernel launch.  Given the large number of such variables, to avoid adding those required by a kernel to its argument list and transferring these to the GPU multiple times, a struct was defined with these as members.  Upon execution, an instance of the struct is initialized, the key/value pairs are read from the config file(s), and the associated struct member associated with that "key" is set to "value".  This struct variable is then transferred to the GPU once and a pointer to its GPU address is added as a parameter to each GPU kernel so this data can be accessed.  Aside from possibly including more parameters in this struct and modifying the GPU functions to accept and access it instead of the const/#define variables, this has been implemented in this commit.  The goal of the next commit is to finish this and completely remove these variables from both host and GPU functions.

To facilitate the distribution of data/images throughout the collaboration, all pCT data will now be stored in a consistent and visually/programmatically transparent format.  This commit includes functions that conform to this format and automatically create the directories/files relevant to adding a new object, scan, run number, preprocessing data, and reconstructed images.  To enforce consistency and avoid duplicating efforts, these same functions should be used and improved by everyone that will generate.  These functions will be made available to everyone, most likely through GitHub to make use of version control.