Config Reading and Struct Definition Complete
Parameter setting through configuration file, the implementation that
parses the file and sets the internal variables, and the struct
definition that contains these parameters is now complete.  An instance
of this struct is now a global variable which is transferred to the GPU
immediately after config reading is complete and the associated GPU
functions now contain an argument for a pointer to where this parameter
container resides on the device so parameters can be accessed by each
GPU function without requiring the struct variable to be transferred to
the GPU on each kernel launch.

The next step is to modify GPU calls to previous #define/const variables
to call the corresponding members of the struct instead.