Completion of configurations struct definition and  routine for specifying options/parameters in external text file
All variables relevant to options/parameters used to control program
behavior have been identified and configurations struct now contains
members for each of these.  The structure of the text file used to
specify these by key/value pair and the functions to read this file and
set the struct member associated with each key to the specified value
have been completed and tested.   The modifiable options/parameters
calculated from other values specified in the configuration file can no
longer be specified explicitly in the file to prevent a conflict between
calculated and specified values.  Once the configuration file has been
read and all key/value pairs used to set their associated struct
members, the struct variable is transferred to the GPU.

The function used to set these values was separated into individual
functions for character array, integer, floating point, and boolean
members because the if/else if/else statement was too long to include
these different types simultaneously and this separation scheme was the
most obvious.

The program now adheres file naming/format scheme established, both for
expected input data naming/location and creation of appropriately
named/located directories where output data is to be written.  The last
step to complete the separation of preprocessing and reconstruction into
2 independently executable stages with preprocessing data imported from
disk and file format/naming scheme enforced is to replace calls to
#define/const variables on GPU by struct members.