Added functions for maintaining a log of the preprocessing/reconstruction executions that have been ran stored in lexicographical order
- Preprocessing and reconstruction runs now write the corresponding
object, scan type, run date, run number, projection data date,
preprocessing date, and reconstruction data.
- The existing log is scanned to maintain lexicographical ordering of
log entries with the order of the above list of parameters defining the
order they appear on each entry line and together specify the path to
where this data is stored.
- If an execution only generates preprocessing data, the reconstruction
date entry is left empty and a subsequent execution using this
preprocessing data then fills in the reconstruction date field as
opposed to generating a new log entry.
- The scan of existing log file detects complete/incomplete matches to
log entry that would be written and whether this is a complete match,
what the log entry # is, and whether either the preprocessing or
reconstruction dates are empty is stored in integer vector.
- Scanning for matches can be used to determine conflicts between config
options/parameters such as log entry exists with same parameters up to
and including preprocessing date but overwriting preprocessing data is
not permitted.
- Detection of all possible conflicts has not been fully implemented
yet, but will be added as needed.