Added ImageJ and Matlab tools to repository
- Completed 1st draft of ImageJ analysis macros to perform insert region
statistics, extract/plot line profiless across diameter of inserts,
apply Sobel filter to image and extract gradient profile along same
insert diameter line as profile was acquired, plot mean RSP vs predicted
RSP, extract slice of image that was analyzed, and generate PNGs for
each of the plots/images and write their underlying data to disk so
external programs (such as Matlab) can import this data to perform
additional analysis.
- Perform this process for the FBP image and each iteration of
reconstruction
- After performing this for each iteration, accumulate (1) the mean RSP
and (2) the discrepancy between the mean observed RSP and predicted RSP
and plot these as a function of iteration, showing the convergence and
error rates.  Again, this data is also written to disk so it it is
possible for another program to import and use this data.
- In addition to generating a PNG for the slice of the image that was
analyzed, the ImageJ analysis macros can optionally extract each slice
of an image and write these to separate PNGs, both for FBP and each
iteration of image reconstruction.  This generates a large number of
images ans to keep the output folder less cluttered, when this is
performed these images are stored in separate folders "FBP_slices" and
"x_slices".
- The Matlab program "ImageJ_Analysis.m" recognizes the format of the
.csv files written by ImageJ and can import, plot, and save the plots as
PNGs.  ImageJ has few plot appearance options and this is sufficient for
analysis purposes, but the Matlab script is intended to be used to
generate better looking images for inclusion in publications.
- Many other Matlab programs have also been added to this commit.  For
example, there are scripts to generate WEPL distribution plots, to
perform filtering and generate filtered image PNGs, converting text
images to PNGs, etc.
- The commit also includes several Matlab functions which perform many
of the common tasks for PCT image analysis and data/image importation.