Added "settings.cfg" file with key/value pairs used to change parameter values and control program behavior without recompiling code
- The key/value pairs in the configuration file "settings.cfg"
correspond to internal program variables
- A struct named "configurations" is defined in the program which has a
member corresponding to each of the keys contained in the configuration
file.
- Command line arguments can be passed to the program which specify the
location of the configuration file if it is not in its default location
and a series of key/value pairs can also be passed which overwrite the
key/value pairs in the configuration file and generate a new
configuration file in cases where the program is to be executed multiple
times with different key/value pairs.
- Thus, command line arguments are parsed first so a modified
configuration file can be generated if desired and then the
configuration file is read and used to set the internal
parameters/options.
- Once the configuration file is read and internal struct members set,
the entire struct object containing all program parameters/options is
transferred to the GPU so these can be accessed by the GPU without
having to manually pass individual parameters/options repeatedly
throughout the program execution.  Instead, tje struct object
persistenly resides on the GPU and only a pointer to its location in GPU
memory need be passed to GPU kernels, greatly simplifying the process
and preventing unnecessary repeated transfers prior to each kernel
invocation.