Improved IO file path/name assignment procedures
- Added extensions to handle IO data/image extensions based on their
specification as TEXT/BINARY in header.
- Added function to define hull, FBP, x_0, x_k, and x file names based
on which, if any, filters are applied to these images.  This includes
addition of "combo" to name if both median and average filter are
applied and their names also include specification of the radius of the
filter neighborhood used in each case.
- Added configuration keys for specifying whether a median and/or
average filter should be applied to each image and the radius of the
filter neighborhood to use in each case.
- Added configuration keys for name of person performing
preprocessing/reconstruction, specifying whether to create a data log
entry for this execution, to redirect console output (stdout) to disk,
if user input requests should be skipped (allowing automatic
script/batch executions), and options for halting execution early and
for turning debug/status console output on/off were moved from header to
configurations file.