Found/added WED calculations from backup code
Found an old version of code that had functions for calculating WED and
this has been added back to the current version of the code.  It is not
clear what stage of development of WED this corresponds to or if it is
complete and works.

In addition, because the naming of global variables and constants in the
current version of code have changed significantly since then, all
global variables and constants/parameters missing from the current
version of the code were copied from the header corresponding to this
old version of the code rather than attempting to change the
variables/parameters the WED functions call to the new names.  Although
long term some of these variables can be replaced with their
corresponding variables from the current version of code, many of these
variables are specific to the phantom used as the target for the WED
calculations and it will not necessarily have the same
dimensions/parameters as the reconstruction image and in these
situations these variables/parameters will remain separate from
reconstruction variables/parameters.

It is already clear that these WED calculations make use of the old
version of the voxel walk algorithm which has known numerical error
issues and these should be replaced at some point, but if the code works
and generates accurate results, this may be postponed until later.