Repaired WED functions to reproduce previous results
The recovered version of the code containing the WED calculation
functions did not work properly, causing the program to crash.  This
issue was resolved and the program now functions as it did, now within
the newest version of the code, and returns the same WED results as it
did before.

Some of the functions, in particular the voxel walk routine that was
adapted to calculate WED to a position within a target volume, were
later found to suffer from accumulation of numerical error, particularly
as the distance travelled along a path increases.  This has little to no
effect for horizontal/vertical paths and preliminary results using these
WED results to place the Bragg peak at a particular point inside a
target volume appear to be accurate, but the results for other angles
have not been verified yet and these are the angles much more likely to
contain errors.  These errors will eventually occur and must be fixed
even if all of the preliminary results prove accurate.

The recovered WED calculation implementation was made hastily and not
only contains the problematic voxel walk approach, but the IO is very
inefficient and the logic is difficult to follow.  Since these WED
calculations need to exist in a reliable and self contained program that
can easily be used by others, these issues must first be addressed.

First, the potentially errant voxel walk algorithm needs to be updated
and the programs logic needs to be cleaned up and made more transparent
and easily understood (including adding comments).  In addition, by
requiring the RSPs of the target volume be specified in a single binary
format file, the entire object can quickly be read into memory in a
single operation instead of the slow and lengthy process using iostream
functions.

Once this has been completed, the next thing is to begin the voxel walk
from the target point outward towards the image edge instead of
calculating the point where the path intersects reconstruction volume
cylinder and then the hull of the object (1st nonzero voxel along path)
before continuing to the target point.  Voxels outside the object should
have 0 RSP and if they do not, they should be considered in the WED
calculation.  Thus, there is no need to find either of these
intersection points if calculations begin on a walk moving outwards from
the target point and the potential errors in these intersection point
calculations and convoluted method for conditionally beginning/ending
the walk can be avoided.

These modifications will not only improve the efficiency and accuracy of
the procedure, but it will also significantly reduce its length and
complexity.  This will make the logic easier to understand and follow,
thereby making it easier for others to familiarize themselves and use
the program.