Improved target volume IO
Combined the images from each slice of target volume (phantom) into a
single text file and a single binary file and added functions for import
these using a single read operation, reading the entire file as a single
block of memory and placing it into the program's internal array used to
store this image.  This significantly reduces the execution time and
simplifies the program.