Modified WED functions to use updated voxel walk algorithm
Removed old voxel walk code from WED functions and combined the two
instances of it with a single instance of new voxel walk algorithm by
preventing WED contributions from voxels outside the object which have
an RSP = 0.001087.  Then the walk from the recon cylinder to the hull
and walk from hull to target were able to be combined. By duplicating
the 2D/3D step functions used in the voxel walk algorithm and modifying
these to calculate the chord length as well, the function that
calculates the WED for each point was dramatically reduced in length and
complexity and it is now much easier to read and understand the logic.
The combination of this, the IO improvements, and other modifications,
the computation time was reduced from 8-10 minutes to ~1-2 seconds for
the ~60 total target points from the 3 different beam angles.

The next commit will attempt to automate the process by which the points
from each beam angle are read by finding all target point files and
reading their beam angle from the file name so the beam angles need not
be specified explicitly and the program does not change, eliminating the
need to change the code each time more/less/different angles are desired
and making it possible for the program to automatically generate results
for every angle by simply double clicking the program's .exe.