Added automated beam angle data search
The program now searches the directory and finds any files whose
filename matches "proxal_distal_XXX_deg.txt" and determines the beam
angle XXX by parsing the filename.  Once all such files and beam angles
have been found, WED calculations are performed on each of the targets
in each of these files and the results are written to disk.

The last step in making the program completely automated is determining
the directory where the target coordinate and target volume files are
located.  This will be inferred from the directory in which the program
is executed or specified explicitly by the user as a command line
execution parameter.  Then all parameters/variables are either constant
or automatically determined at run time, so the program's source code
need not be modified or recompiled to perform WED analysis for
more/less/different beam angles and/or beam targets.  Thus, the program
can be distributed as a .exe and users need only create text files for
each desired beam angle containing the desired target coordinates for
that angle, name and place these appropriately, and double click the
.exe to calculate and write the WED results to disk.  The focus for the
next commit is implementing this procedure.