FInished 1st attempt at parallel DROP imeplementation w/ precalculated MLP
Preprocessed data is written to disk and imported into memory for image
reconstruction.  This preprocessed data includes MLP, WEPL, voxels per
MLP, chord length, and history data (all hull entry/exit
coordinates/angles, hull entry voxels in x/y/z directions, gantry angle,
and bin numbers) as well as the hull and initial iterate images.  All of
this data except for history data is written to disk in the same order
in which histories are used in image reconstruction, with this order
defined using a variation of the golden section search algorithm and is
stored in the history_sequence array.

There are 2 options for importing this preprocessed data: (1) each files
can be imported in its entirety into host memory and the required data
for each DROP block can be copies from the vectors and transferred to
the GPU or (2) the files can be left open and the data for each DROP
block can be imported directly into arrays and transferred to the GPU as
needed.

There have been 2 data files as part of the preprocessed data required
for image reconstruction, the # of voxels per MLP and the effective
chord length for each history.  Since each MLP contains a variable # of
intersected voxels, this was written to disk before each MLP itself so
this could be read and used to determine the # of bytes to read to
import the following MLP.  This required 2 file read operations per MLP
as the size of each had to be determined before reading it.  By writing
these lengths to a separate file, this information can be read into
memory and then used to determine the # of bytes to read to import any
number of MLPs using a single read operation.

The MLP for each history in the DROP block is then placed in a single
array and transferred to the GPU and each thread (history) must access
the appropriate portion of the array containing its MLP.  The # of
voxels per MLP specifies the length of each MLP and summing these for
each preceding MLP provides the index where each MLP begins, specifying
the offset index and range of indices where each MLP resides.  The DROP
update calculations require inner product vales and these must be
completed before calculating the update contribution from each history.
The inner products and addition of these contributions from each history
to the x_update array for each voxel in its MLP are performed in
separate loops, providing the opportunity to copy the MLP from global to
shared memory duing the 1st loop so it can be accessed more quickly
during the 2nd loop.  This option will be explored and its performance
compared to the default approach which accesses the MLP from global
memory during both loops.  It must also be determined if there is enough
shared memory to store the MLP for each history and if there is a DROP
block size where this is no longer possible.  Such a limit will depend
on compute system and the data set and will even vary between blocks, so
if efficiency is not markedly improved, developing a procedure for
determining the block size limit and conditionally exploit shared memory
may not be worthwhile.

To improve image reconstruction we have adopted usage of the effective
chord length approach developed by Scott Penfold and this is calculated
using various combinations of hull entry/exit coordinates/angles.  The
history data file containing this information is quite large (10+ GB)
and unlike most preprocessed data, it is not written to disk in the
order histories are used in reconstruction.  Thus, the entire file must
be read into memory and the history sequence used to identify and
extract the required data for each history in a DROP block one at a
time.  The minimum data required for these calculations is the hull
entry/exit xy/xz angles but even if this data was written to a separate
file, this would still require this entire file to be read into memory
and extracted 1 element at a time before transferring this data into 4
separate arrays on the GPU.  However, by calculating the effective chord
length during MLP, it can be written to its own file in the appropriate
reconstruction order so it can be imported and transferred into a single
array on the GPU using the same procedure as is used for WEPL and voxels
per MLP.

Voxels per MLP and effective chord length have always been generated as
needed and limited to local scope, but this approach could not be
retained in the parallel implementation without great difficulty and
significant performance costs. The voxels per MLP data is effectively a
requirement since storing the MLP for each history of a DROP block in a
single array on the GPU is the only feasible option and a thread must
know where its MLP resides within it.  Precalculating a mundane value
like effective chord length is not ideal but this has the lowest
computational/storage cost.  Together, these improve the ease and
efficiency of the preprocessed data transfers to/from the disk/GPU
required to separate MLP from the image reconstruction process.