Fixed config file read and script based control of config parameter values
When the configurations struct was added, it was used to replace calls
to #define/const variables and to prevent naming conflicts with the
#define/const variables and configurations struct members, "_D was added
to the end of each struct member name.  Once all calls to the
#define/const variables has been replaced with calls to the struct
members, these were deleted and the _D removed from their names.  A
global replace of _D resulted in numerous other variable name/value
changes and this caused the IO directory naming procedure to crash.  The
actual issue turned out to be the char* array "pCT_Data" changing to
"pCTata" and the program trying to access this directory, but this
simple mistake was difficult to find.

It was also found that the procedure that imports the config file,
changes parameter values, and outputs the modified config file was
writing an extra line at the end of the file.  The 2nd time this happens
the vector would contain gibberish and write this to the file.  This did
not damage the program, but it sent this gibberish to the console window
and it was removed.