Added parallel DROP implementation for single block scenarios using precalculated MLP
When DROP can be performed using a single GPU block, thread
synchronization can be enforced using the syncthreads() functions.
Thus, there is no need to separate the update calculations and the
application of these updates to the image into individual kernels.  In
addition, the MLP paths and the initial image x can be transferred to
shared memory and the x_update and S arrays can be created, stored, and
updated in shared memory.  Then the image residing in global memory can
be updated from the values of these shared memory arrays.  This commit
includes these additions.

Since the precalculated preprocessing data is reasonably sized, except
for the MLP path data, their entire files can be imported into host
memory and transferred to the GPU in entirety as well.  It may be
possible in some cases to also import the entire MLP path file into host
memory, but this is unlikely to fit on the GPU in any realistic
situations, so this information must be transferred to the GPU as
needed.  The MLP data that is transferred to the GPU is either imported
from disk or directly from host memory if this data was able to be
imported in its entirety into host memory.

When the preprocessed data is read into memory in its entirety,
reconstruction can proceed by transferring all of this to the GPU or as
needed for each DROP block.  If it can only be read iteratively, it is
only transferred to the GPU iteratively.  For the MLP data, it can
either be read into host memory in its entirety or iteratively as
needed, but in both cases, this data is transferred to the GPU as needed
since it will not fit on the GPU at once (in any useful situation).
Each of these scenarios are supported and controlled via configuration
file options/parameters.

Much of the code was also organized better for this commit and the
series of tasks performed immediately after execution, such as applying
any command line arguments and reading the program parameters/options
from the config file, were placed in a single configure_program()
function.

In addition, since the experimental data tends to contain significantly
more histories with negative WEPL, many with extremely negative WEPL
(-100 to -1000), an additional struct member was created and added as a
key to the configuration routine which is used to define a lower
threshold for hull detection below which a history is ignored and not
used for hull detection.  Thus, a proton is considered to have missed
the object entirely if its WEPL is in the range between the lower and
upper limit and both of these limits can be specified in the
configuration file.