1st finished implementation of pCT data automated soft link organizational script
Run date is now parsed from folder name, the for loop over angles now
runs as many times as it needs to depending on the angle interval.  In
addition, the file names are split into tokens using _ as delimiter and
the run# folder can sometimes have _inf or _sup etc. after run_xxx.  The
previous version would support 1 additional term separated by '_' but
now this will work for any # of additional '_' separated tokens.  Links
are also given the global path so the links can be moved as desired
without breaking the links.