Finished and verified MLP lookup table procedure
The procedure for generating/importing the (1) sin/cos (2) scattering
coefficient and (3) scattering matrix polynomial tables and performed
tests of MLP using these tables which verified the implementation was
valid.  Several versions of MLP were created, each with 1 additional
improvement added so debugging of these routines could be performed more
easily.  The reconstructed images for these versions of MLP are visually
identical but their values do vary slightly and in rare cases fairly
significantly in voxels near the edge of the object.  The 1st new
version of MLP did not include use of any of the tables, this
improvement was confined to elimination of multiplications by 0/1 and
unnecessary temporary variables and algebraic manipulations to simplify
calculations.  Although some of these simplifications result in a
mismatch between variable names and the actual calculations performed.
The simplified calculations make the code more difficult to read and
understand, but it also reduces the # of lines of code in both the setup
calculations and each loop of MLP, so together the  modified MLP is
actually easier to read and understand.  There is documentation showing
the derivation of the closed form equations and algebraic manipulations
of these equations to simplify the implementation.

It was predicted that the MLP with lookup tables would be ~3X-4X faster
than the original, but the tests show that the sequential
implementations of MLP is 7X-8X faster, which suggested there was
something wrong but it is now clear that the reconstructed images are
very close to that from the original implementation so the only
explanation is that the approximation computational cost of
sin/cos/exp/log functions were significantly lower than reality.

It is still unclear why there is such a large difference (up to 0.6)
between the original version of MLP and the full lookup table
implementation for about 10 voxels in the image and these voxels are
always on or close to the boundary of the object, so further detailed
tests need to be performed on the various versions of MLP to hopefully
determine if this error is from a particular lookup table or more than 1
and which one(s) and hopefully find a way to avoid this.  Although the
maximum difference between the original MLP and the 1st new
implementation with just algebraic manipulations is 0.00125 and most are
much smaller, many or even most voxels are at least slightly different
and this was not expected, it remains to be seen if the version with
unnecessary calculations is actually more accurate as it should be given
that this presumably results in less error and fewer opportunities for
error.  The step sizes for each of the tables have been decreased, 10X,
20X, and 2X smaller for (1), (2), and (3), respectively.  This should
reduce error and this hopefully eliminates the sporadic issues with
voxel errors.

There are also a few more improvements that can be made.  The
calculations involving +=/-= etc of a constant floating point value are
going to be performed explicitly using a counter, e.g.,

u_1 += MLP_U_STEP -> u_1 += n * MLP_U_STEP

Since MLP now uses lookup tables, it is very important that all other
values are as accurate as possible to prevent error propagation.  When
the same floating point values is repeatedly added to a variable, the
result is rounded to double precision each time and after several
applications of this, the rounding/truncation error can accumulate in
the final values.

The GPU implementation of image reconstruction now needs to be modified
to use the new version of MLP and the data transfers and GPU
configurations for the MLP endpoints and DROP reconstruction procedures.