Finished 1st draft of GPU image reconstruction testing routine
The various functions with different transfer modes and efficiency
improvements which need to be tested can now be chosen based on the
values of enum control variables.  Based on these values, multiple
combinations of these approaches can be tested with the same code by
using a script to specify which are desired.  I the previous commit
there were numerous functions which performed nearly the exact same
operations but were defined as separate functions so that the function
names would correspond directly, even though this resulted in several
identical functions with simply different names so it was clear which
functions would use it.  This was intended to make things easier but it
had the opposite effect.  Thus, the code was modified so that common
routines were made more generic so several functions could use the same
helper functions.