Fixed issue causing incorrect # for reconstruction_histories
The wrong # for reconstruction_histories was caused by an issue with
MSC_edge_detection which resulted in random and sporadic "dots" in the
hull image, i.e., randomly distributed individual voxels outside the
actual hull object with value 1 which should have had value 0.  The
average filter applied to the hull to expand its boundary, which is the
combination of an average filter and thresholding routine to limit
possible values to 0/1, resulted in small squares with side lengths
HULL_FILTER_RADIUS occurring where these individual voxels were
initially located.  Then when the MLP endpoints were determined, the
protons that hit these randomly distributed squares and were then
identified as having hit the hull when they actually didn't.  Worst of
all, the location and # of these dots varies between executions so the
reconstruction_histories varies significantly as well.  Howeer, by
preventing these "dots" from appearing in the first place then prevents
a varying number of reconstruction_histories as well.

The MSC_edge_detection_GPU kernel was modified so that instead of just
calculating the difference in counts between neighboring MSC_counts
voxels, it now also sets all voxels with MSC_counts values greater than
MSC_DIFF_THRESH are also set to 0.  Ultimately this is not a great
approach because the baseline # of counts for voxels inside the object
increases as the total # of histories increases, so this approach may
require changes to MSC_DIFF_THRESH depending on the data set, though
this probably does not need to change often or by much in most, perhaps
even all, cases.